add feature tests for guests
- refactored some other feature tests.
