+++ app/views/sessions/new.html.erb
@@ -10,5 +10,6 @@
 
     <%= f.submit "Login" %>
     <% end  %>
+    <%= link_to "Register", signup_path %>
   </div>
 </div>
