+++ spec/features/user_must_login_to_apply_for_jobs_spec.rb
@@ -1,21 +1,25 @@
 require "rails_helper"
-RSpec.feature "UserMustLoginToCheckout", type: :feature do
+RSpec.feature "User", type: :feature do
   context "when logged out" do
 
   let!(:company) { Company.create!(name: "acme",
                                    information: "asdf") }
 
-  let!(:job) { company.jobs.create!(title: "dev",
+  let!(:job) { company.jobs.create!(title: "Job 1",
                                     description: "Job 1 description") }
+  let!(:job2) { company.jobs.create!(title: "Job 2",
+                                     description: "Job 2 description") }
 
-    it "redirects the user to a login page if they are not logged in" do
+    it "is redirected to a login page" do
       visit company_job_path(job.company, job)
-
       click_button "Add to Favorites"
+      visit company_job_path(job2.company, job2)
+      click_button 'Add to Favorites'
 
       visit cart_path
 
-      expect(page).to have_content("dev")
+      expect(page).to have_content('Job 1')
+      expect(page).to have_content('Job 2')
 
       click_button "Apply to Jobs"
