+++ spec/features/visitor_can_view_jobs_spec.rb
@@ -17,11 +17,5 @@ RSpec.feature "VisitorCanViewJobs", type: :feature do
       expect(page).to have_content("Jr Dev")
       expect(page).to have_content("Denver, CO")
     end
-
-    it "can view all existing jobs" do
-      visit jobs_path
-
-      expect(current_path).to eq(jobs_path)
-    end
   end
 end
